+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features / Enhancements
 
 - Add Disabled Text Area element type (#243)
+- Add Confirmation Window options (#242)
 
 ## 3.2.1 (2023-09-07)